+++ lilith/src/main/java/de/huxhorn/lilith/services/clipboard/ClipboardFormatter.java
@@ -1,6 +1,6 @@
 /*
  * Lilith - a log event viewer.
- * Copyright (C) 2007-2016 Joern Huxhorn
+ * Copyright (C) 2007-2017 Joern Huxhorn
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,6 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
+
 package de.huxhorn.lilith.services.clipboard;
 
 import de.huxhorn.lilith.services.BasicFormatter;
@@ -44,6 +45,11 @@ public interface ClipboardFormatter
 	 * @return the accelerator of this formatter, can be null.
 	 */
 	String getAccelerator();
+
+	default Integer getMnemonic()
+	{
+		return null;
+	}
 
 	/**
 	 * Returns true, if this is a native Lilith formatter, i.e. the formatter is part