+++ lilith/src/main/java/de/huxhorn/lilith/swing/menu/FocusMenu.java
@@ -1,6 +1,6 @@
 /*
  * Lilith - a log event viewer.
- * Copyright (C) 2007-2016 Joern Huxhorn
+ * Copyright (C) 2007-2017 Joern Huxhorn
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,9 +15,12 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
+
 package de.huxhorn.lilith.swing.menu;
 
+import de.huxhorn.lilith.swing.AbstractLilithAction;
 import de.huxhorn.lilith.swing.ApplicationPreferences;
+import de.huxhorn.lilith.swing.LilithActionId;
 import de.huxhorn.lilith.swing.actions.FilterAction;
 import de.huxhorn.lilith.swing.actions.FocusCallLocationAction;
 import de.huxhorn.lilith.swing.actions.FocusFormattedMessageAction;
@@ -30,6 +33,8 @@ import de.huxhorn.lilith.swing.actions.FocusThreadGroupNameAction;
 import de.huxhorn.lilith.swing.actions.FocusThreadNameAction;
 import de.huxhorn.lilith.swing.actions.FocusThrowableAction;
 import de.huxhorn.lilith.swing.actions.FocusThrowablesAction;
+import java.awt.event.ActionEvent;
+import javax.swing.Action;
 import javax.swing.JMenuItem;
 
 public class FocusMenu
@@ -39,7 +44,7 @@ public class FocusMenu
 
 	public FocusMenu(ApplicationPreferences applicationPreferences, boolean htmlTooltip)
 	{
-		super("Focus", applicationPreferences, htmlTooltip);
+		super(ACTION_INSTANCE, applicationPreferences, htmlTooltip);
 		createUI();
 		viewContainerUpdated();
 	}
@@ -163,5 +168,22 @@ public class FocusMenu
 		registerAccessComponent(responseHeaderMenu);
 		registerAccessComponent(null);
 		registerAccessComponent(remoteUserItem);
+	}
+
+	private static final Action ACTION_INSTANCE=new MenuAction();
+
+	private static class MenuAction
+			extends AbstractLilithAction
+	{
+		private static final long serialVersionUID = -6989836389361332927L;
+
+		MenuAction()
+		{
+			super(LilithActionId.FOCUS);
+		}
+
+		public void actionPerformed(ActionEvent e)
+		{
+		}
 	}
 }